LayoutTests/platform/efl/fast/forms/validationMessage-expected.txt

Test for validationMessage DOM property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


input patternMismatch: Please match the requested format:
input valueMissing: Please fill out this field.
textarea valueMissing: Please fill out this field.
select valueMissing: Please select an item in the list.
input typeMismatch: Please enter an email address
input badInput: Please enter a number.
badInput and valueMissing:
PASS numberInput.validationMessage is nonRequiredBadInputMessage
PASS but.validationMessage is ''
PASS anoninput.validationMessage is ''
PASS happyFieldset.validationMessage is ''
PASS happySelect.validationMessage is ''
PASS happyOutput.validationMessage is ''
PASS happyObject.validationMessage is ''
PASS happyKeygen.validationMessage is ''
PASS successfullyParsed is true

TEST COMPLETE
